+++ src/common/schedulers/__init__.py
@@ -26,11 +26,12 @@ class Job:
     """Class with submission specifications
     """
 
-    def __init__(self, job_script, job_dir, account=None, env_file=None, additional_options=None):
+    def __init__(self, job_script, job_dir, account=None, job_env=None,
+                 additional_options=None):
         self.job_dir = job_dir
         self.job_script = job_script
         self.account = account
-        self.env_file = env_file
+        self.job_env = job_env
         self.opts = additional_options if additional_options else []
 
 
@@ -47,7 +48,7 @@ class JobScript:
         dependency_id=None,
         account=None,
         constraint=None,
-        env_file=None
+        job_env=None
     ):
         self.name = name
         self.time = time
@@ -56,7 +57,7 @@ class JobScript:
         self.dependency_id = dependency_id
         self.account = account
         self.constraint = constraint
-        self.env_file = env_file
+        self.job_env = job_env
 
 
 class JobScheduler(abc.ABC):